requirements.txt
numpy==1.19.2
tqdm==4.46.0
matplotlib==3.1.1
seaborn==0.11.0
timm==0.3.2
watchdog==0.9.0
scipy==1.4.1
tabulate==0.8.7
torch==1.7.1
pandas==1.0.5
graphviz==0.13
torchvision==0.8.2
ipython==7.19.0
Pillow==8.1.0
protobuf==3.14.0
typing==3.7.4.3
